fix: add browser-safe svg entry
