+++ package.json
@@ -40,6 +40,7 @@
     "@types/node": "^25.2.2",
     "@types/ws": "^8.18.1",
     "@vitest/coverage-v8": "4.1.0",
+    "esbuild": "0.27.7",
     "eslint": "^9.0.0",
     "eslint-config-prettier": "^10.0.0",
     "eslint-plugin-import-x": "^4.16.2",
