+++ packages/core/tsup.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from "tsup";
 
 export default defineConfig({
-  entry: ["src/index.ts"],
+  entry: ["src/index.ts", "src/browser.ts"],
   format: ["cjs", "esm"],
   dts: { resolve: ["@pptx-glimpse/renderer"] },
   clean: true,
